energy_consumption/requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0

# Prophet and forecasting
prophet>=1.1.4
holidays>=0.29

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.11.0

# Machine learning and evaluation
scikit-learn>=1.1.0

# Data handling
pytz>=2022.1
python-dateutil>=2.8.0

# Configuration
PyYAML>=6.0

# Optional: for better performance
cmdstanpy>=1.0.0

# Development
jupyter>=1.0.0
notebook>=6.4.0